refact change the project structure as a whole
I was dissatisfied with the current state of the project structure. It was basically a spaghetti (still is, at some points), so I refactored it to be more concise and loose coupled. I'm trying to make everything depend on the domain layer, as it is the most important layer, because every other one can be replaced, besides this one (unless I want to rewrite the application logic).

Plus, I'm reading Dependency Injections - Principles, Practices and Patterns, which is helping me to get a better grasp on how to design a software. I still have some more things to figure out, but this change already is a big leap towards a better direction.